+++ nodes/blynk-ws-out-write.js
@@ -44,12 +44,25 @@ module.exports = function(RED) {
 					text: "blynk-ws-out-write.status.disconnected"
 				});
 			});
+			this.blynkClient.on("disabled", function() {
+				node.status({
+					fill: "red",
+					shape: "dot",
+					text: "blynk-ws-out-write.status.disabled"
+				});
+			});
 		} else {
 			this.error(RED._("blynk-ws-out-write.errors.missing-conf"));
 		}
 		this.on("input", function(msg) {
-			if (msg.hasOwnProperty("payload") && node.blynkClient && node.blynkClient.logged) {
-				var payload = Buffer.isBuffer(msg.payload) ? msg.payload : RED.util.ensureString(msg.payload);
+
+			//no input operation if client not connected or disabled
+			if(!node.blynkClient || !node.blynkClient.logged) {
+				return; 
+			}
+				
+			if (msg.hasOwnProperty("payload")) {
+				var payload = Array.isArray(msg.payload) ? msg.payload : RED.util.ensureString(msg.payload);
 				var pin = node.pin;
 				if(node.pinmode == 1) {
 					if (!msg.hasOwnProperty("pin")) {